+++ src/features/ChatInput/RuntimeConfig/useWorkingTreeStatus.ts
@@ -1,7 +1,7 @@
 import { isDesktop } from '@lobechat/const';
 
 import { useClientDataSWR } from '@/libs/swr';
-import { electronSystemService } from '@/services/electron/system';
+import { electronGitService } from '@/services/electron/git';
 
 /**
  * Working-tree dirty-file breakdown for the current cwd.
@@ -12,7 +12,7 @@ import { electronSystemService } from '@/services/electron/system';
 export const useWorkingTreeStatus = (dirPath?: string) => {
   const key = isDesktop && dirPath ? ['git-working-tree-status', dirPath] : null;
 
-  return useClientDataSWR(key, () => electronSystemService.getGitWorkingTreeStatus(dirPath!), {
+  return useClientDataSWR(key, () => electronGitService.getGitWorkingTreeStatus(dirPath!), {
     focusThrottleInterval: 5 * 1000,
     revalidateOnFocus: true,
     shouldRetryOnError: false,
